Fixed issue with invalid serialization of array when response contains empty object. Fixed issue #277 Rely on #292
