+++ src/JMS/Serializer/JsonSerializationVisitor.php
@@ -54,6 +54,11 @@ class JsonSerializationVisitor extends GenericSerializationVisitor
     {
         $result = parent::visitArray($data, $type, $context);
 
+        if ($result instanceof \ArrayObject) {
+            $result = $result->getArrayCopy();
+            $this->setRoot($result);
+        }
+
         if (null !== $this->getRoot() && isset($type['params'][1]) && 0 === count($result)) {
             // ArrayObject is specially treated by the json_encode function and
             // serialized to { } while a mere array would be serialized to [].
